Update clean up python code
