--- server/exec.py
@@ -1,4 +1,0 @@
-from subprocess import call
-
-def buildDocker(imageName, filePath):
-    call(["docker", "build", "-t", "autocontain/" + imageName, filePath])
